Fix test_pluck_entity_data_species_identity to use explicit dtypes for cross-platform compatibility
Fixed issue #42 by specifying explicit dtypes (int64 and float64) for the numpy arrays used in the test. This resolves dtype mismatches between Windows and macOS platforms.

+++ src/tests/test_net_utils.py
@@ -60,12 +60,14 @@ def test_validate_graph_attributes(sbml_dfs):
 def test_pluck_entity_data_species_identity(sbml_dfs):
     # Take first 10 species IDs
     species_ids = sbml_dfs.species.index[:10]
-    # Create mock data
+    # Create mock data with explicit dtype to ensure cross-platform consistency
+    # Fix for issue-42: Use explicit dtypes to avoid platform-specific dtype differences
+    # between Windows (int32) and macOS/Linux (int64)
     mock_df = pd.DataFrame({
         'string_col': [f"str_{i}" for i in range(10)],
-        'mixed_col': np.arange(-5, 5),
-        'ones_col': np.ones(10),
-        'squared_col': np.arange(10),
+        'mixed_col': np.arange(-5, 5, dtype=np.int64),  # Explicitly use int64
+        'ones_col': np.ones(10, dtype=np.float64),      # Explicitly use float64
+        'squared_col': np.arange(10, dtype=np.int64),   # Explicitly use int64
     }, index=species_ids)
     # Assign to species_data
     sbml_dfs.species_data["mock_table"] = mock_df
